Add files via upload
Many changes to the CSS and Javascript of the website, header is nearly complete, look for the rest of the website is planned out at this point. The forecast predicts I'll have the header perfected lunchtime tomorrow and will then begin placing and stying the other pages while removing all content I transfer from the home page (i.e. the oil slider)